Custom HitBox Editor In Development
Huge Updates made to integrate hitbox data into the animation system.

The framework is 90% complete on that front... I think.

Still need to develop the PrepareFrame method. That should be it.

On the other hand, the custom hitbox editor, which is designed to make the authoring of hitbox data easy and fast, is 80% complete. The entire framework of this tool has been devised and mostly implemented.

Up next:
- Finish some remaining data modification functionality in the hitbox editor.

- Finish the code that exports hitbox editor data back to the scriptable object file, and that should be it for the hitbox editor.