Essential HitBox Editor Quality Updates
Now when you select a hitbox in the scene, the corresponding HitBoxEditor element in the hitbox list will highlight green so that you know which element is responsible for the selection's values.

Additionally, now when a hitbox has a key frame that corresponds with the current frame, it's HitBoxEditor element will highlight orange. This allows the user to know whether they are or are not on a key frame, and it lets them have a way to know whether a key frame even exists on a given hitbox.

Up Next: Need to implement the saving functionality to send data from the editor back to the file being edited.