HitBoxEditor "Completed"
Can now save data modified within the HitBox Editor back to the FightAnimation file!

HitBox editor is essentially complete. The rest of the features that relied on it may now be developed!

A few minor updates to the HitBoxEditor will likely come in the near future, plus bug fixes, as I can already foresee some likely bugs in interpretation of the data in-game versus in-editor due to differing character animation setups.