HitBox Editor Refactor
That small bug that I predicted a couple commits ago turned out to be a two-night refactor of the hit box editor. Essentially, now hitboxes are childed to the character, and all their data is now local to the character transform.

Now that that's been fixed, the hitbox system is in great shape. Hitboxes are now appearing in the right location and scale on the player character!