Hit Interaction System Development
Last update, hitbox pipeline of developing hitbox animations in animation editor was completed, with hitboxes correctly appearing when moves are executed after being authored in the hitbox editor.

Now, I have begun development of the hit detection system between players through these hitboxes. Baby steps:

--Hitboxes are now triggers
--Hitboxes for each player are placed in their own layer in the Unity collision matrix, so each player's hitboxes only interact with the enemy player's hitboxes and not each other.
-- The hitboxes' gameObjects' tags now signify whether they are a hurt or hitbox.
--With these tags, hurtboxes only respond to contact from hitboxes.

--Collision detection with these constraints is now working. Collision event only outputs debug log.

Next up: somehow funnel this collision detection in an organized way into the rest of the game systems so that the game systems can appropriately respond to collisions... Not sure how I am going to do this... As I have to deal with Unity's not-so-great built-in architecture for the calling of collision functions in the back-end.